requirements.txt
PyYAML
launchpadlib
click

# https://github.com/python/importlib_resources/pull/255
importlib-resources>=5.9

git+https://git.launchpad.net/~freyes/python-apt@setup-requires#egg=python-apt
git+https://github.com/openstack-charmers/charmed-openstack-info.git#egg=charmed-openstack-info
git+https://github.com/freyes/charmhub-lp-tools.git@use-pbr#egg=charmhub-lp-tools
GitPython
